+++ packages/commons/src/events/FieldTypeMapping.ts
@@ -160,7 +160,7 @@ export function mapFieldTypeToZod(field: FieldConfig, actionType?: ActionType) {
     case FieldType.VERIFICATION_STATUS:
     case FieldType.ID:
     case FieldType.LOADER:
-    case FieldType.HIDDEN:
+    case FieldType.ALPHA_HIDDEN:
       schema = field.required ? NonEmptyTextValue : TextValue
       break
     case FieldType.NUMBER:
@@ -252,7 +252,7 @@ export function mapFieldTypeToEmptyValue(field: FieldConfig) {
     case FieldType.QR_READER:
     case FieldType.ID_READER:
     case FieldType.LOADER:
-    case FieldType.HIDDEN:
+    case FieldType.ALPHA_HIDDEN:
       return null
     case FieldType.ADDRESS:
       return {
@@ -562,7 +562,7 @@ export const isHiddenFieldType = (field: {
   config: FieldConfig
   value: FieldValue | FieldUpdateValue
 }): field is { value: undefined; config: HiddenField } => {
-  return field.config.type === FieldType.HIDDEN
+  return field.config.type === FieldType.ALPHA_HIDDEN
 }
 
 export type NonInteractiveFieldType =