+++ packages/commons/src/events/test.utils.ts
@@ -187,7 +187,7 @@ function mapFieldTypeToMockValue(
     case FieldType.OFFICE:
     case FieldType.LINK_BUTTON:
     case FieldType.LOADER:
-    case FieldType.HIDDEN:
+    case FieldType.ALPHA_HIDDEN:
       return `${field.id}-${field.type}-${i}`
     case FieldType.VERIFICATION_STATUS:
       return 'verified'