+++ CHANGELOG.md
@@ -18,6 +18,7 @@ HTTP input now accepts `field('..')` references in the HTTP body definition.
 - Refactor the tRPC context to allow defining public procedures that don't require authentication.
 - Remove legacy mongo migration status outputs and skip typecheck which reduced the migration service startup time by 66%.
 - The postgres migration files now get restored to their original state (i.e. without the environment variables being replaced) regardless of the migration passing or not
+- Added experimental ALPHA_HIDDEN form field type, allowing configurable default/derived values and conditional inclusion in form submissions.
 
 ## 1.9.2